Fixed compilation warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3040 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/ffmpeg_in/ffmpeg_demux.c
@@ -460,7 +460,7 @@ static int ff_url_read(void *h, unsigned char *buf, int size)
 
 	full_size = 0;
 	if (ffd->buffer_used) {
-		if (ffd->buffer_used>size) {
+		if (ffd->buffer_used > (u32) size) {
 			ffd->buffer_used-=size;
 			memcpy(ffd->buffer, ffd->buffer+size, sizeof(char)*ffd->buffer_used);
 #ifdef FFMPEG_DUMP_REMOTE
